+++ client/package.json
@@ -16,12 +16,14 @@
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
-    "reactstrap": "^8.6.0",
-    "ts-node": "^9.0.0"
+    "reactstrap": "^8.6.0"
   },
   "scripts": {
-    "start": "concurrently \"ts-node ./service/index.ts\" \"cd client && react-scripts start\"",
-    "dev": "ts-node ./service/index.ts"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom --passWithNoTests",
+    "eject": "react-scripts eject",
+    "lint": "eslint src/**/*.ts src/**/*.tsx"
   },
   "devDependencies": {
     "@types/express": "^4.17.8",
@@ -43,5 +45,18 @@
     "nodemon": "^2.0.4",
     "prettier": "2.1.2",
     "typescript": "^4.0.3"
-  }
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "proxy": "http://localhost:5000"
 }
