+++ client/src/components/Home/index.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import Home from "./";
+import Home from ".";
 
 test("renders main app page", () => {
   const { getByText } = render(<Home />);
