+++ client/src/components/InfoSpinner/index.tsx
