+++ client/src/components/InfoSpinner/infospinner.css
