+++ client/src/components/NavBar/Profile.tsx
